# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SchoolCore - Multi-tenant school management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and get session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /schools/{schoolId}/classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /schools/{schoolId}/classes/{classId}:
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

  # ─── SUBJECTS ────────────────────────────────────────────────────────────────
  /schools/{schoolId}/subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"

  # ─── ATTENDANCE ───────────────────────────────────────────────────────────────
  /schools/{schoolId}/attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: submitAttendance
      tags: [attendance]
      summary: Submit attendance for a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Attendance submitted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /schools/{schoolId}/attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Attendance summary per class/student
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"

  # ─── RESULTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          description: Phone number with country code e.g. +233241234567
    OtpResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    OtpVerify:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AuthSession:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, phone, role]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, school_admin, teacher, parent]
        schoolId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── SCHOOLS ───────────────────────────────────────────────────────────────
...
    Class:
      type: object
      required: [id, name, schoolId]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: ["string", "null"]
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        studentCount:
          type: integer
        schoolId:
          type: integer
    ClassInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        level:
          type: string
        teacherId:
          type: ["integer", "null"]
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
        teacherId:
          type: ["integer", "null"]

    # ─── SUBJECTS ──────────────────────────────────────────────────────────────
    Subject:
      type: object
      required: [id, name, schoolId]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        schoolId:
          type: integer
    SubjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: ["string", "null"]

    # ─── ATTENDANCE ────────────────────────────────────────────────────────────
    AttendanceRecord:
      type: object
      required: [id, studentId, classId, date, status]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late]
        schoolId:
          type: integer
    AttendanceInput:
      type: object
      required: [classId, date, records]
      properties:
        classId:
          type: integer
        date:
          type: string
          format: date
        records:
          type: array
          items:
            type: object
            properties:
              studentId:
                type: integer
              status:
                type: string
                enum: [present, absent, late]
            required: [studentId, status]
    AttendanceSummary:
      type: object
      required: [studentId, studentName, presentCount, absentCount, lateCount, percentage]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        className:
          type: string
        presentCount:
          type: integer
        absentCount:
          type: integer
        lateCount:
          type: integer
        percentage:
          type: number
...